Add tests mixing pizzas
These are all happy path tests demonstrating how left and right pizzas
might be mixed. The obvious missing case here is without any parens,
however the code won't compile in this case as it cannot tell which
order to apply everything.

Essentially these tests just check that a paren expression on either
side of a left pizza works as expected.